Move core code out of docs/viewer
