--- docs/viewer/test.ts
@@ -1,9 +1,0 @@
-import { test } from "./test/util/cappucino";
-
-import * as ComposeAlphaTest from "./test/compose-alpha-test"
-import * as ExampleFileTest from "./test/example-file-test"
-
-ComposeAlphaTest
-ExampleFileTest
-
-test();
