+++ src/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "ifcx",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "build-viewer": "esbuild viewer/render.ts --bundle --outfile=../docs/viewer/render.mjs --external:three --format=esm",
+    "test": "esbuild test/test.ts --bundle --outfile=test.js --external:three --platform=node && node test.js",
+    "pkg-cli": "esbuild ifcx-cli/ifcx-cli.ts --bundle --outfile=ifcx-cli.js --external:three --platform=node && pkg -t node*-win-x64 --output ifcx.exe ./ifcx-cli.js",
+    "serve": "serve ."
+  },
+  "author": "",
+  "description": "",
+  "devDependencies": {
+    "@types/node": "^22.13.10",
+    "@types/three": "^0.171.0",
+    "chai": "^5.2.0",
+    "esbuild": "^0.24.2",
+    "glob": "^11.0.1",
+    "pkg": "^5.8.1",
+    "serve": "^14.2.4"
+  },
+  "dependencies": {
+    "three": "^0.171.0"
+  }
+}
