+++ src/test/test.ts
@@ -1,0 +1,9 @@
+import { test } from "./util/cappucino";
+
+import * as ComposeAlphaTest from "./compose-alpha-test"
+import * as ExampleFileTest from "./example-file-test"
+
+ComposeAlphaTest
+ExampleFileTest
+
+test();
