+++ src/viewer/composed-object.ts
